group page rework / modal fixes
