+++ components/bank/modals/sendModal.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import SendBox from '../components/sendBox';
 import { CombinedBalanceInfo } from '@/utils/types';
 import { useEffect } from 'react';
+import { createPortal } from 'react-dom';
 
 interface SendModalProps {
   modalId: string;
@@ -13,6 +14,9 @@ interface SendModalProps {
   refetchHistory: () => void;
   selectedDenom?: string;
   setOpen?: (isOpen: boolean) => void;
+  isGroup?: boolean;
+  admin?: string;
+  refetchProposals?: () => void;
 }
 
 export default function SendModal({
@@ -25,6 +29,9 @@ export default function SendModal({
   selectedDenom,
   isOpen,
   setOpen,
+  isGroup,
+  admin,
+  refetchProposals,
 }: SendModalProps) {
   const handleClose = () => {
     if (setOpen) {
@@ -44,14 +51,30 @@ export default function SendModal({
     return () => document.removeEventListener('keydown', handleEscape);
   }, [isOpen]);
 
-  return (
+  const modalContent = (
     <dialog
       id={modalId}
-      className={`modal ${isOpen ? 'modal-open' : ''} z-[999]`}
+      className={`modal ${isOpen ? 'modal-open' : ''}`}
       onClose={handleClose}
+      style={{
+        position: 'fixed',
+        top: 0,
+        left: 0,
+        right: 0,
+        bottom: 0,
+        zIndex: 9999,
+        backgroundColor: 'transparent',
+        padding: 0,
+        margin: 0,
+        height: '100vh',
+        width: '100vw',
+        display: isOpen ? 'flex' : 'none',
+        alignItems: 'center',
+        justifyContent: 'center',
+      }}
     >
       <div
-        className="modal-box max-w-xl mx-auto rounded-[24px] bg-[#F4F4FF] dark:bg-[#1D192D] shadow-lg relative z-[1000]"
+        className="modal-box max-w-xl mx-auto rounded-[24px] bg-[#F4F4FF] dark:bg-[#1D192D] shadow-lg relative"
         aria-label="send modal"
       >
         <form method="dialog">
@@ -72,11 +95,33 @@ export default function SendModal({
           refetchBalances={refetchBalances}
           refetchHistory={refetchHistory}
           selectedDenom={selectedDenom}
+          isGroup={isGroup}
+          admin={admin}
+          refetchProposals={refetchProposals}
         />
       </div>
-      <form method="dialog" className="modal-backdrop">
+      <form
+        method="dialog"
+        className="modal-backdrop"
+        style={{
+          position: 'fixed',
+          top: 0,
+          left: 0,
+          right: 0,
+          bottom: 0,
+          zIndex: -1,
+          backgroundColor: 'rgba(0, 0, 0, 0.3)',
+        }}
+      >
         <button onClick={handleClose}>close</button>
       </form>
     </dialog>
   );
+
+  // Only render if we're in the browser
+  if (typeof document !== 'undefined') {
+    return createPortal(modalContent, document.body);
+  }
+
+  return null;
 }
