+++ components/groups/components/index.tsx
@@ -1,5 +1,5 @@
 export * from './CountdownTimer';
 export * from '../modals/groupInfo';
-export * from './groupProposals';
+export * from './groupControls';
 export * from './myGroups';
 export * from '../../react/StepIndicator';
